Update translation and add a few languages that are close to completely translated

+++ main/misc/fetchtranslations.sh
@@ -15,7 +15,10 @@ wget -nv https://crowdin.com/backend/download/project/ics-openvpn.zip
 # Chinese language require zh-CN and zh-TW
 
 typeset -A langhash
-langhash=(zh-CN zh-rCN zh-TW zh-rTW id-ID in ca-ES ca cs-CZ cs et-EE et ja-JP ja ko-KR ko sv-SE sv uk-UA uk vi-VN vi sl-SI sl da-DK da be-BY be)
+langhash=(zh-CN zh-rCN zh-TW zh-rTW id-ID in ca-ES ca cs-CZ cs et-EE et
+          ja-JP ja ko-KR ko sv-SE sv uk-UA uk vi-VN vi sl-SI sl da-DK
+          da be-BY be he-IL he ar-SA ar fa-IR fa si-LK si pt-BR pt-rBR
+          sr-SP sr-rSP el-GR el sk-SK sk)
 
 langtoinclude="de es fr hu it no nl pl pt ro ru tr"
 
@@ -34,4 +37,6 @@ do
     tar -xv -C src/ui/res/values-$rlang/ --strip-components 2 -f ics-openvpn.zip res/values-$alang/
 done
 
-rm ics-openvpn.zip
+if [ "$NODELETE" = "" ]; then
+  rm ics-openvpn.zip
+fi
